✨ refactor Home component; integrate Header, Footer, and ModalGuestInvite components, and streamline guest input handling ✨ add ModalGuestInvite component; implement email invitation functionality with toast notifications for duplicates ✨ add Footer component; include terms of service and privacy policy links ✨ add Header component; display logo and invitation prompt fix import path for Home component in App
