+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import Home from "./pages/Home"
+import Home from "./pages/Home/Home"
 
 export default function App() {
   return (
