+++ frontend/src/components/ModalConfirmationTrip/ModalConfirmationTrip.tsx
@@ -1,0 +1,5 @@
+export default function ModalConfirmationTrip(){
+  return(
+    
+  )
+}
