+++ frontend/src/pages/Footer/Footer.tsx
@@ -1,0 +1,16 @@
+export default function Footer() {
+  return (
+    <p className="text-zinc-500 text-sm">
+      En planifiant votre voyage par l'intermédiaire de PlannTrip, vous acceptez
+      automatiquement <br /> à nos{' '}
+      <a href="" className="text-zinc-300 underline">
+        conditions générales d'utilisation
+      </a>{' '}
+      et à notre{' '}
+      <a href="" className="text-zinc-300 underline">
+        politique de confidentialité
+      </a>
+      .
+    </p>
+  )
+}
