+++ frontend/src/pages/Home/Home.tsx
@@ -1,0 +1,40 @@
+import { useState } from 'react'
+import SearchForm from '../../components/SearchForm/SearchForm'
+import GuestForm from '../../components/GuestForm/GuestForm'
+import ModalGuestInvite from '../../components/ModalGuestInvite/ModalGuestInvite'
+import Footer from '../Footer/Footer'
+import Header from '../Header/Header'
+
+export default function Home() {
+  const [inputOpen, setInputOpen] = useState(false)
+  const [guestModalOpen, setGuestModalOpen] = useState(false)
+  const [emailsToInvite, setEmailsToInvite] = useState<string[]>([])
+
+  const openGuestInput = () => setInputOpen(true)
+  const closeGuestInput = () => setInputOpen(false)
+  const openGuestModal = () => setGuestModalOpen(true)
+  const closeGuestModal = () => setGuestModalOpen(false)
+
+  return (
+    <div className="max-w-3xl mx-auto px-6 text-center space-y-10 ">
+      <Header />
+      <div className="space-y-4">
+        <SearchForm
+          openGuestInput={openGuestInput}
+          inputOpen={inputOpen}
+          closeGuestInput={closeGuestInput}
+        />
+        {inputOpen && <GuestForm openGuestModal={openGuestModal} guestsCount={emailsToInvite.length}/>}
+      </div>
+      <Footer />
+
+      {guestModalOpen && (
+        <ModalGuestInvite
+          closeGuestModal={closeGuestModal}
+          emailsToInvite={emailsToInvite}
+          setEmailsToInvite={setEmailsToInvite}
+        />
+      )}
+    </div>
+  )
+}
